Conflicts:
	client/code/style.css

+++ server/code/persistence_manager.js
@@ -19,12 +19,14 @@ var redis = require("redis");
 var Farmer = require('./models/farmer');
 var Tile = require('./models/tile');
 var GameState = require('./models/gamestate');
+var StoredCrop = require('./models/storedCrop');
 
-var PersistenceManager = {
+module.exports = {
 	asyncblock: null,
 	keys: {
 		databaseVersion: null,
 		lastPersistDate: null,
+		storedCropsPrefix: null,
 		farmersPrefix: null,
 		settingsPrefix: null,
 		tickRate: null,
@@ -37,6 +39,7 @@ var PersistenceManager = {
 		boardSizeY: null,
 		boardTilesPrefix: null
 	},
+	enabled: true,
 	databaseVersion: 1,
 	client: null,
 	onError: function(err) {
@@ -47,9 +50,28 @@ var PersistenceManager = {
 			console.log("PersistenceManager - Error while persisting : " + err);
 		}
 	},
-	persist: function(callback) {
+	clear: function() {
+		this.enabled = false;
+		console.log("PersistenceManager - CLEARING GAME DATA and disabling PM for this session");
+		console.log("PersistenceManager - Please reboot the server, it will now exit");
 		this.asyncblock((function(flow) {
+			this.client.flushdb(flow.add());
+			flow.wait();
+			process.exit(1337);
+		}).bind(this));
+	},
+	persist: function(callback) {
+		if(!this.enabled) {
+			console.log("PersistenceManager - Persistence failed : PM Disabled");
 			return;
+		}
+		this.asyncblock((function(flow) {
+			// Quick, dirty, non exhaustive and maybe not even working sanity check
+			if(GameState == undefined || GameState == null || GameState.board == undefined || GameState.board == null
+				|| GameState.board.size.x == undefined || GameState.board.size.x == null || isNaN(GameState.board.size.x)) {
+				console.log("PersistenceManager - Bad GameState, bailing out");
+				return;
+			}
 			console.log("PersistenceManager - Persisting gamestate");
 			var startDate = Date.now();
 			this.client.flushdb(flow.add());
@@ -61,6 +83,11 @@ var PersistenceManager = {
 				// Key : farmer:<nickname>
 				this.client.hmset(this.keys.farmersPrefix + farmer.nickname, farmer.getPersistable(), flow.add());
 			}).bind(this));
+			var tmpStoredCrop;
+			for(var key in GameState.board.storedCrops) {
+				tmpStoredCrop = GameState.board.storedCrops[key];
+				this.client.hmset(this.keys.storedCropsPrefix + tmpStoredCrop.id, tmpStoredCrop.getPersistable(), flow.add());
+			}
 			GameState.board.tiles.forEach((function(tileLine) {
 				tileLine.forEach((function(tile) {
 					// Key : board:tile:<y>:<x>
@@ -78,8 +105,11 @@ var PersistenceManager = {
 		}).bind(this), callback);
 	},
 	load: function(callback) {
-		this.asyncblock((function(flow) {
+		if(!this.enabled) {
+			console.log("PersistenceManager - Loading failed : PM Disabled");
 			return;
+		}
+		this.asyncblock((function(flow) {
 			console.log("PersistenceManager - Loading gamestate");
 			var startDate = Date.now();
 			// This syntax allows me to get the database value in a synchronous way
@@ -100,6 +130,7 @@ var PersistenceManager = {
 			this.client.get(this.keys.lastPersistDate, flow.set(this.keys.lastPersistDate));
 			this.client.keys(this.keys.boardTilesPrefix + "*", flow.set('tilesKeys'));
 			this.client.keys(this.keys.farmersPrefix + "*", flow.set('farmersKeys'));
+			this.client.keys(this.keys.storedCropsPrefix + "*", flow.set('storedCropsKeys'));
 			GameState.settings.tickRate = parseInt(flow.get(this.keys.tickRate));
 			GameState.settings.startMoney = parseInt(flow.get(this.keys.startMoney));
 			GameState.board.size.x = parseInt(flow.get(this.keys.boardSizeX));
@@ -107,6 +138,7 @@ var PersistenceManager = {
 			GameState.lastPersistDate = parseInt(flow.get(this.keys.lastPersistDate));
 			var farmersKeys = flow.get('farmersKeys');
 			var tilesKeys = flow.get('tilesKeys');
+			var storedCropsKeys = flow.get('storedCropsKeys');
 			farmersKeys.forEach((function(key) {
 				this.client.hgetall(key, flow.add(key));
 			}).bind(this));
@@ -115,6 +147,10 @@ var PersistenceManager = {
 				this.client.hgetall(key, flow.add(key));
 			}).bind(this));
 			var dbTiles = flow.wait();
+			storedCropsKeys.forEach((function(key) {
+				this.client.hgetall(key, flow.add(key));
+			}).bind(this));
+			var dbStoredCrops = flow.wait();
 
 			var farmers = {};
 			GameState.farmers = [];
@@ -133,6 +169,9 @@ var PersistenceManager = {
 				var tmpFarmer = farmers[farmer.nickname];
 				tmpFarmer.email = farmer.email;
 				tmpFarmer.password = farmer.password;
+				tmpFarmer.password = farmer.password;
+				tmpFarmer.health = parseInt(farmer.health);
+				tmpFarmer.admin = farmer.admin == "true" ? true : false;
 				tmpFarmer.last_pos.x = parseInt(farmer.last_pos_x);
 				tmpFarmer.last_pos.y = parseInt(farmer.last_pos_y);
 				tmpFarmer.money = parseInt(farmer.money);
@@ -146,9 +185,11 @@ var PersistenceManager = {
 					// We consider that the database is consistent and the farmers exist.
 					tmpFarmer.allied_farmers.push(farmers[key]);
 				}).bind(this));
+				tmpFarmer.inventory = JSON.parse(farmer.inventory);
 			}
 
 			GameState.board.tiles = [];
+			var y; // It's also used later
 			for(y = 0; y < GameState.board.size.y; y++) {
 				GameState.board.tiles[y] = [];
 			}
@@ -161,6 +202,10 @@ var PersistenceManager = {
 				tmpTile.fertility = parseFloat(tile.fertility);
 				tmpTile.max_fertility = parseFloat(tile.max_fertility);
 				tmpTile.maturity = parseFloat(tile.maturity);
+				tmpTile.growingCrop = JSON.parse(tile.growingCrop);
+				// Don't forget to fix it later
+				tmpTile.storedCrops = JSON.parse(tile.storedCrops);
+
 				if(tile.owner != "dummy") {
 					// We consider that the database is consistent bla bla
 					tmpTile.owner = farmers[tile.owner];
@@ -176,11 +221,56 @@ var PersistenceManager = {
 				GameState.board.tiles[tmpTile.position.y][tmpTile.position.x] = tmpTile;
 			}
 
+			// Read the stored crops
+			for(var key in dbStoredCrops) {
+				var storedCrop = dbStoredCrops[key];
+				var tmpStoredCrop = new StoredCrop(GameState.settings.crops[storedCrop.crop]
+					,farmers[storedCrop.owner]
+					,0);
+				tmpStoredCrop.id = storedCrop.id;
+				tmpStoredCrop.harvested_quantity = storedCrop.harvested_quantity;
+				tmpStoredCrop.time_left = storedCrop.time_left;
+				if(storedCrop.parent_tile == "null") {
+					tmpStoredCrop.parent_tile = null;
+				} else {
+					var tilePosition = JSON.parse(storedCrop.parent_tile);
+					tmpStoredCrop.parent_tile = GameState.board.tiles[tilePosition.y][tilePosition.x];
+				}
+				GameState.board.storedCrops[tmpStoredCrop.id] = tmpStoredCrop;
+			}
+
+			// Set the real stored crops in the tiles
+			var tmpTile;
+			var tmpStoredCrops;
+			for(y = 0; y < GameState.board.size.y; y++) {
+				for(var x = 0; x < GameState.board.size.x; x++) {
+					tmpTile = GameState.board.tiles[y][x];
+					tmpStoredCrops = tmpTile.storedCrops;
+					tmpTile.storedCrops = [];
+					tmpStoredCrops.forEach(function (storedCropId) {
+						tmpTile.storedCrops.push(GameState.board.storedCrops[storedCropId]);
+					});
+				}
+			}
+
+			// Set the real stored crops in the inventories
+			var tmpFarmer;
+			for(var i = 0; i < GameState.farmers.length; i++) {
+				tmpFarmer = GameState.farmers[i];
+				tmpStoredCrops = tmpFarmer.inventory;
+				tmpFarmer.inventory = [];
+				tmpStoredCrops.forEach(function (storedCropId) {
+					tmpFarmer.inventory.push(GameState.board.storedCrops[storedCropId]);
+				});
+			}
+
 			console.log("PersistenceManager - Loading done in " + (Date.now() - startDate) + " ms");
 			return true;
 		}).bind(this), callback);
 	}
 };
+
+var PersistenceManager = module.exports;
 
 // Populate keys here. We can't do this before because some values depend on others.
 PersistenceManager.keys.databaseVersion = "databaseVersion";
@@ -196,6 +286,7 @@ PersistenceManager.keys.boardPrefix = "board:";
 PersistenceManager.keys.boardSizeX = PersistenceManager.keys.boardPrefix + "size:x";
 PersistenceManager.keys.boardSizeY = PersistenceManager.keys.boardPrefix + "size:y";
 PersistenceManager.keys.boardTilesPrefix = PersistenceManager.keys.boardPrefix + "tile:";
+PersistenceManager.keys.storedCropsPrefix = PersistenceManager.keys.boardPrefix + "stored_crop:";
 
 
 // Synchronous node.js, deal with it.
